standardising boosts to Double

+++ elastic4s-core/src/main/scala/com/sksamuel/elastic4s/searches/queries/QueryStringQueryDefinition.scala
@@ -12,7 +12,7 @@ case class QueryStringQueryDefinition(query: String,
                                       defaultOperator: Option[String] = None,
                                       defaultField: Option[String] = None,
                                       enablePositionIncrements: Option[Boolean] = None,
-                                      fields: Seq[(String, Float)] = Nil,
+                                      fields: Seq[(String, Double)] = Nil,
                                       fuzziness: Option[String] = None,
                                       fuzzyMaxExpansions: Option[Int] = None,
                                       fuzzyPrefixLength: Option[Int] = None,
@@ -37,7 +37,7 @@ case class QueryStringQueryDefinition(query: String,
   def defaultOperator(op: String): QueryStringQueryDefinition = copy(defaultOperator = op.some)
   def operator(op: String): QueryStringQueryDefinition = defaultOperator(op)
 
-  def asfields(fields: String*): QueryStringQueryDefinition = copy(fields = fields.map(f => (f, -1F)))
+  def asfields(fields: String*): QueryStringQueryDefinition = copy(fields = fields.map(f => (f, -1D)))
 
   def splitOnWhitespace(splitOnWhitespace: Boolean): QueryStringQueryDefinition =
     copy(splitOnWhitespace = splitOnWhitespace.some)
@@ -76,10 +76,10 @@ case class QueryStringQueryDefinition(query: String,
     copy(quoteFieldSuffix = quoteFieldSuffix.some)
 
   def field(name: String): QueryStringQueryDefinition =
-    copy(fields = fields :+ (name, -1F))
+    copy(fields = fields :+ (name, -1D))
 
   def field(name: String, boost: Double): QueryStringQueryDefinition =
-    copy(fields = fields :+ (name, boost.toFloat))
+    copy(fields = fields :+ (name, boost.toDouble))
 
   def defaultField(field: String): QueryStringQueryDefinition =
     copy(defaultField = field.some)
