+++ elastic4s-core/src/main/scala/com/sksamuel/elastic4s/searches/queries/matches/MatchAllQueryDefinition.scala
@@ -2,10 +2,10 @@ package com.sksamuel.elastic4s.searches.queries.matches
 
 import com.sksamuel.elastic4s.searches.queries.QueryDefinition
 
-case class MatchAllQueryDefinition(boost: Option[Float] = None,
+case class MatchAllQueryDefinition(boost: Option[Double] = None,
                                    queryName: Option[String] = None) extends QueryDefinition {
-  def boost(boost: Float): MatchAllQueryDefinition = copy(boost = Option(boost))
-  def withBoost(boost: Float): MatchAllQueryDefinition = copy(boost = Option(boost))
+  def boost(boost: Double): MatchAllQueryDefinition = copy(boost = Option(boost))
+  def withBoost(boost: Double): MatchAllQueryDefinition = copy(boost = Option(boost))
   def queryName(queryName: String): MatchAllQueryDefinition = copy(queryName = Option(queryName))
   def withQueryName(queryName: String): MatchAllQueryDefinition = copy(queryName = Option(queryName))
 }
